[CloverLeaf] Make scripts consistent with TeaLeaf

+++ cloverleaf/skl-swan/benchmark.sh
@@ -27,22 +27,10 @@ COMPILER=${2:-$DEFAULT_COMPILER}
 SCRIPT=`realpath $0`
 SCRIPT_DIR=`realpath $(dirname $SCRIPT)`
 
-export BENCHMARK_PLATFORM=bdw-swan
 export BENCHMARK_EXE=clover_leaf
+export CONFIG="skl"_"$COMPILER"
 export SRC_DIR=$PWD/CloverLeaf_ref
-export RUN_DIR=$PWD/CloverLeaf-$BENCHMARK_PLATFORM-$COMPILER
-
-# Fetch source code if necessary
-if [ ! -r "$SRC_DIR/clover.f90" ]
-then
-    if ! "$SCRIPT_DIR/../fetch.sh"
-    then
-        echo
-        echo "Failed to fetch source code."
-        echo
-        exit 1
-    fi
-fi
+export RUN_DIR=$PWD/CloverLeaf-$CONFIG
 
 
 # Set up the environment
@@ -78,6 +66,15 @@ esac
 # Handle actions
 if [ "$ACTION" == "build" ]
 then
+    # Fetch source code
+    if ! "$SCRIPT_DIR/../fetch.sh"
+    then
+        echo
+        echo "Failed to fetch source code."
+        echo
+        exit 1
+    fi
+
     # Perform build
     rm -f $SRC_DIR/$BENCHMARK_EXE $RUN_DIR/$BENCHMARK_EXE
     if ! eval make -C $SRC_DIR -B $MAKE_OPTS
@@ -93,8 +90,16 @@ then
 
 elif [ "$ACTION" == "run" ]
 then
+    if [ ! -x "$RUN_DIR/$BENCHMARK_EXE" ]
+    then
+        echo "Executable '$RUN_DIR/$BENCHMARK_EXE' not found."
+        echo "Use the 'build' action first."
+        exit 1
+    fi
+
     qsub $SCRIPT_DIR/run.job \
-        -o CloverLeaf-$BENCHMARK_PLATFORM-$COMPILER.out \
+        -d $RUN_DIR
+        -o CloverLeaf-$CONFIG.out \
         -N cloverleaf \
         -V
 else
